Time: 3 ms (41.69%), Space: 49.2 MB (21.8%) - LeetHub

+++ 0713-subarray-product-less-than-k/0713-subarray-product-less-than-k.java
@@ -1,20 +1,20 @@
 class Solution {
     public int numSubarrayProductLessThanK(int[] nums, int k) {
-        if (k <= 1) return 0; 
+        if (k <= 1) return 0;
 
         int low = 0;
         int count = 0;
-        long product = 1;
+        int product = 1;  
 
-        for (int high = 0; high < nums.length; high++) {  
+        for (int high = 0; high < nums.length; high++) {
             product *= nums[high];
 
-            while (product >= k) {  
+            while (product >= k) {
                 product /= nums[low];
                 low++;
             }
 
-            count += (high - low + 1);  
+            count += (high - low + 1);
         }
 
         return count;
